requirements.txt
setuptools>=68.0.0
numpy>=1.19.0
pybullet==3.2.6
opencv-python>=4.5.0
opencv_contrib_python==4.10.0.84
customtkinter==5.2.2
matplotlib>=3.3.0
scipy>=1.10.0
scikit-image>=0.18.0
spyder==5.4.3
spyder-kernels==2.4.3
argcomplete==3.5.1
bitstring==4.2.3
cryptography==41.0.0
ecdsa==0.19.0
intelhex==2.3.0
Pillow==10.2.0
pyaudio==0.2.14
pyserial==3.5
pytest==7.3.1
PyYAML==6.0.1
reedsolo==1.7.0
requests==2.31
sphinx==4.5.0
sherpa-ncnn>=1.8.0
sentencepiece>=0.1.99
protobuf>=3.20.1
cpm_kernels>=1.0.11
openai==1.55.3
httpx==0.27.2
tqdm==4.65.0
langchain>=0.2.17
langchain-community>=0.2.19
langchain-openai>=0.1.25
freetype-py==2.4.0
vtk==9.3.0
pyzed
xacrodoc==0.7.0
reportlab
wmi==1.5.1; sys_platform == "win32"
python-can
sounddevice